+++ server/src/main/java/org/elasticsearch/index/fielddata/TriangleTreeReader.java
@@ -1,20 +1,21 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.spatial.index.fielddata;
+package org.elasticsearch.index.fielddata;
 
 import org.elasticsearch.common.io.stream.ByteArrayStreamInput;
 
 import java.io.IOException;
 
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeWriter.LEFT;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeWriter.LINE;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeWriter.POINT;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeWriter.RIGHT;
+import static org.elasticsearch.index.fielddata.TriangleTreeWriter.LEFT;
+import static org.elasticsearch.index.fielddata.TriangleTreeWriter.LINE;
+import static org.elasticsearch.index.fielddata.TriangleTreeWriter.POINT;
+import static org.elasticsearch.index.fielddata.TriangleTreeWriter.RIGHT;
 
 /**
  * A tree reader for a previous serialized {@link org.elasticsearch.geometry.Geometry} using
