+++ server/src/main/java/org/elasticsearch/index/fielddata/TriangleTreeVisitor.java
@@ -1,15 +1,16 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.spatial.index.fielddata;
+package org.elasticsearch.index.fielddata;
 
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeWriter.AB_FROM_TRIANGLE;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeWriter.BC_FROM_TRIANGLE;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeWriter.CA_FROM_TRIANGLE;
+import static org.elasticsearch.index.fielddata.TriangleTreeWriter.AB_FROM_TRIANGLE;
+import static org.elasticsearch.index.fielddata.TriangleTreeWriter.BC_FROM_TRIANGLE;
+import static org.elasticsearch.index.fielddata.TriangleTreeWriter.CA_FROM_TRIANGLE;
 
 /** Visitor for triangle interval tree.
  *
