+++ server/src/main/java/org/elasticsearch/index/fielddata/TriangleTreeWriter.java
@@ -1,11 +1,12 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
 
-package org.elasticsearch.xpack.spatial.index.fielddata;
+package org.elasticsearch.index.fielddata;
 
 import org.apache.lucene.document.ShapeField;
 import org.apache.lucene.index.IndexableField;
@@ -22,7 +23,7 @@ import java.util.List;
  * This is a tree-writer that serializes a list of {@link ShapeField.DecodedTriangle} as an interval tree
  * into a byte array.
  */
-class TriangleTreeWriter {
+public class TriangleTreeWriter {
 
     static final byte LEFT = 1;
     static final byte RIGHT = 1 << 1;
