+++ server/src/main/java/org/elasticsearch/index/mapper/CartesianShapeIndexer.java
@@ -1,10 +1,11 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
  * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
  */
-package org.elasticsearch.xpack.spatial.index.mapper;
+package org.elasticsearch.index.mapper;
 
 import org.apache.lucene.document.XYShape;
 import org.apache.lucene.index.IndexableField;
@@ -21,7 +22,6 @@ import org.elasticsearch.geometry.MultiPolygon;
 import org.elasticsearch.geometry.Point;
 import org.elasticsearch.geometry.Polygon;
 import org.elasticsearch.geometry.Rectangle;
-import org.elasticsearch.index.mapper.ShapeIndexer;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -47,7 +47,7 @@ public class CartesianShapeIndexer implements ShapeIndexer {
         return visitor.fields;
     }
 
-    private class LuceneGeometryVisitor implements GeometryVisitor<Void, RuntimeException> {
+    private static class LuceneGeometryVisitor implements GeometryVisitor<Void, RuntimeException> {
         private List<IndexableField> fields = new ArrayList<>();
         private String name;
