+++ x-pack/plugin/spatial/src/main/java/org/elasticsearch/xpack/spatial/index/fielddata/GeoShapeScriptDocValues.java
@@ -10,6 +10,9 @@ package org.elasticsearch.xpack.spatial.index.fielddata;
 import org.apache.lucene.index.IndexableField;
 import org.elasticsearch.common.geo.Orientation;
 import org.elasticsearch.geometry.Geometry;
+import org.elasticsearch.index.fielddata.CentroidCalculator;
+import org.elasticsearch.index.fielddata.CoordinateEncoder;
+import org.elasticsearch.index.fielddata.GeometryDocValueWriter;
 import org.elasticsearch.index.mapper.GeoShapeIndexer;
 import org.elasticsearch.script.GeometryFieldScript;
 import org.elasticsearch.search.aggregations.support.ValuesSourceType;
