+++ x-pack/plugin/spatial/src/main/java/org/elasticsearch/xpack/spatial/index/fielddata/Tile2DVisitor.java
@@ -7,10 +7,13 @@
 
 package org.elasticsearch.xpack.spatial.index.fielddata;
 
+import org.elasticsearch.index.fielddata.TriangleTreeVisitor;
+import org.elasticsearch.index.fielddata.TriangleTreeWriter;
+
 import static org.apache.lucene.geo.GeoUtils.orient;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeVisitor.abFromTriangle;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeVisitor.bcFromTriangle;
-import static org.elasticsearch.xpack.spatial.index.fielddata.TriangleTreeVisitor.caFromTriangle;
+import static org.elasticsearch.index.fielddata.TriangleTreeVisitor.abFromTriangle;
+import static org.elasticsearch.index.fielddata.TriangleTreeVisitor.bcFromTriangle;
+import static org.elasticsearch.index.fielddata.TriangleTreeVisitor.caFromTriangle;
 
 /**
  * A reusable tree reader visitor for a previous serialized {@link org.elasticsearch.geometry.Geometry} using
